added reentrancy tests to existing vickreyauction test file
