+++ test/VickreyAuction.test.ts
@@ -218,4 +218,84 @@ describe('VickreyAuction', function () {
             await expect(vickreyAuction.connect(bidder1).claim(auctionId)).to.be.revertedWith('Auction has not ended yet');
         });
     });
+
+        async function advanceTime(seconds: number) {
+            await ethers.provider.send('evm_increaseTime', [seconds]);
+            await ethers.provider.send('evm_mine', []);
+       }
+
+   describe('Reentrancy Regression Tests', function () {
+    const BID_COMMIT_DURATION = 1000;
+    const BID_REVEAL_DURATION = 90000;
+    const COMMIT_FEE = ethers.parseEther('0.1');
+    const BID_AMOUNT_1 = ethers.parseEther('10');
+    const BID_AMOUNT_2 = ethers.parseEther('20');
+
+    async function createAuction() {
+      await mockNFT.connect(auctioneer).approve(vickreyAuction.getAddress(), 1);
+      await vickreyAuction.connect(auctioneer).createAuction(
+        'Test NFT',
+        'Test Description',
+        'https://example.com/image.jpg',
+        0,
+        await mockNFT.getAddress(),
+        1,
+        await biddingToken.getAddress(),
+        ethers.parseEther('5'),
+        BID_COMMIT_DURATION,
+        BID_REVEAL_DURATION,
+        COMMIT_FEE
+      );
+      return 0;
+    }
+
+    async function commitBids(auctionId: number) {
+      const salt1 = ethers.id('salt1');
+      const salt2 = ethers.id('salt2');
+
+      const commitment1 = ethers.solidityPackedKeccak256(['uint256', 'bytes32'], [BID_AMOUNT_1, salt1]);
+      const commitment2 = ethers.solidityPackedKeccak256(['uint256', 'bytes32'], [BID_AMOUNT_2, salt2]);
+
+      await vickreyAuction.connect(bidder1).commitBid(auctionId, commitment1, { value: COMMIT_FEE });
+      await vickreyAuction.connect(bidder2).commitBid(auctionId, commitment2, { value: COMMIT_FEE });
+
+      return { salt1, salt2 };
+    }
+
+    it('should update accumulatedCommitFee before external call in revealBid', async function () {
+      const auctionId = await createAuction();
+      const { salt1 } = await commitBids(auctionId);
+
+      await advanceTime(BID_COMMIT_DURATION + 1);
+
+      const before = await vickreyAuction.auctions(auctionId);
+      expect(before.accumulatedCommitFee).to.equal(COMMIT_FEE * 2n);
+      await biddingToken.connect(bidder1).approve(vickreyAuction.getAddress(), BID_AMOUNT_1);
+      await vickreyAuction.connect(bidder1).revealBid(auctionId, BID_AMOUNT_1, salt1);
+
+      const after = await vickreyAuction.auctions(auctionId);
+      expect(after.accumulatedCommitFee).to.equal(COMMIT_FEE);
+    });
+
+    it('should update accumulatedCommitFee to 0 before external call in withdraw', async function () {
+      const auctionId = await createAuction();
+      const { salt1, salt2 } = await commitBids(auctionId);
+
+      await advanceTime(BID_COMMIT_DURATION + 1);
+
+      await biddingToken.connect(bidder1).approve(vickreyAuction.getAddress(), BID_AMOUNT_1);
+      await biddingToken.connect(bidder2).approve(vickreyAuction.getAddress(), BID_AMOUNT_2);
+
+      await vickreyAuction.connect(bidder1).revealBid(auctionId, BID_AMOUNT_1, salt1);
+      await vickreyAuction.connect(bidder2).revealBid(auctionId, BID_AMOUNT_2, salt2);
+
+      await advanceTime(BID_REVEAL_DURATION + 1);
+
+      await vickreyAuction.connect(auctioneer).withdraw(auctionId);
+
+      const after = await vickreyAuction.auctions(auctionId);
+      expect(after.accumulatedCommitFee).to.equal(0);
+    });
+  });
+
 });
